IPE: Drop unused components and fix linux32 smoke test issue

Remove the IPE components that are no longer in use.

Verified compilation on clang/darwin, gcc/linux64 STD, and gcc/linux64 LLVM.
Just to be safe I also ran against release/ on darwin and linux64

+++ compiler/ipe/IpeModule.cpp
@@ -370,11 +370,11 @@ void IpeModule::describeAllModules(int offset)
     *tptr = '\0';
   }
 
-  printf("%sThere are %3lu modules. Global pool %8lu / %8lu\n",
+  printf("%sThere are %3d modules. Global pool %8d / %8d\n",
          pad,
-         sModules.size(),
-         IpeEnv::gMemoryTail,
-         IpeEnv::gMemorySize);
+         (int) sModules.size(),
+         (int) IpeEnv::gMemoryTail,
+         (int) IpeEnv::gMemorySize);
 
   for (size_t i = 0; i < sModules.size(); i++)
   {
